[#1] 이미지 캐싱 시스템 주석 처리
